router organization
