--- backend/routes/routes.js
@@ -1,7 +1,0 @@
-const express = require("express")
-
-const items = require("./items")
-
-const router = express.Router()
-
-router.use('/api/items', items)
